Create `SpeechToText` Library functionality
Keep methods specific to the Speech To Text api in the appropriate
class. Create a method foreach method in the api.

The podcasts should be added in `resources/podcasts`, we can override
this if need be. We may want to consider adding this directory to
the `.gitignore` however I have added two small files for dev purposes.

An example of getting a transcript from a podcast would be:

    $speechToText = new SpeechToText();
    $speechToText->recognize('carlin_blow_it.opus');
    $transcripts = $speechToText->getTranscripts();